box_firmware: convert display code to be task-oriented
