+++ hardware/firmware/box_rp2040/src/display.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <stdint.h>
+
+void display_init(void);
+void display_task(void);
